Add some documentation and rename login to authentication
