+++ frontend/lib/widgets/home.dart
@@ -4,6 +4,7 @@ import 'post.dart';
 import 'timeline.dart';
 import 'users.dart';
 
+/// Widget which provides [Scaffold] and allows accessing other views.
 class Home extends StatefulWidget {
   @override
   State<StatefulWidget> createState() {
@@ -12,15 +13,18 @@ class Home extends StatefulWidget {
 }
 
 class _HomeState extends State<Home> {
+  /// Selected index
   int _index = 0;
 
   final String _title = 'Full Stack Example';
 
+  /// Views which can be shown
   final List<Widget> _views = [
     Timeline(),
     UserList(),
   ];
 
+  /// Sets the index when user selects a different tab
   _onTappedIcon(int index) {
     setState(() {
       this._index = index;
